requirements.txt
numpy==1.23.5
django==4.2.3
djangorestframework==3.14.0
gunicorn==20.1.0
scikit-learn==1.2.2
